+++ visionframework/core/trackers/reid_tracker.py
@@ -8,7 +8,7 @@ from scipy.spatial.distance import cdist
 from .base_tracker import BaseTracker
 from ...data.track import STrack
 from ...data.detection import Detection
-from ..reid import ReIDExtractor
+from ..processors.reid_extractor import ReIDExtractor
 from ...utils.logger import get_logger
 from ...utils.config import Config